+++ src/iad_agrid.c
@@ -29,6 +29,9 @@ static int AGrid_N = 0;
 static int AGrid_Cap = 0;
 static int AGrid_Search = -1;
 
+static double AGrid_target_mr = 0.0;
+static double AGrid_target_mt = 0.0;
+
 static int AGrid_Initialized = 0;
 static double AGrid_slab_n = 0;
 static double AGrid_slab_n_top = 0;
@@ -147,10 +150,11 @@ static int agrid_add_or_get(double a, double b, double g)
     return idx;
 }
 
-static double agrid_raw_interp_error(int c00, int c10, int c01, int c11, int cc)
+static double agrid_raw_interp_error(int c00, int c10, int c01, int c11, int cc, int *brackets)
 {
     double mr00, mt00, mr10, mt10, mr01, mt01, mr11, mt11, mrcc, mtcc;
     double interp_mr, interp_mt;
+    double mr_lo, mr_hi, mt_lo, mt_hi;
     agrid_entry_t *e;
 
     e = &AGrid_Cache[c00];
@@ -163,6 +167,45 @@ static double agrid_raw_interp_error(int c00, int c10, int c01, int c11, int cc)
     abg_sphere_mr_mt(e->a, e->b, e->g, e->ur1, e->ut1, e->uru, e->utu, &mr11, &mt11);
     e = &AGrid_Cache[cc];
     abg_sphere_mr_mt(e->a, e->b, e->g, e->ur1, e->ut1, e->uru, e->utu, &mrcc, &mtcc);
+
+    mr_lo = mr_hi = mr00;
+    if (mr10 < mr_lo)
+        mr_lo = mr10;
+    if (mr10 > mr_hi)
+        mr_hi = mr10;
+    if (mr01 < mr_lo)
+        mr_lo = mr01;
+    if (mr01 > mr_hi)
+        mr_hi = mr01;
+    if (mr11 < mr_lo)
+        mr_lo = mr11;
+    if (mr11 > mr_hi)
+        mr_hi = mr11;
+    if (mrcc < mr_lo)
+        mr_lo = mrcc;
+    if (mrcc > mr_hi)
+        mr_hi = mrcc;
+
+    mt_lo = mt_hi = mt00;
+    if (mt10 < mt_lo)
+        mt_lo = mt10;
+    if (mt10 > mt_hi)
+        mt_hi = mt10;
+    if (mt01 < mt_lo)
+        mt_lo = mt01;
+    if (mt01 > mt_hi)
+        mt_hi = mt01;
+    if (mt11 < mt_lo)
+        mt_lo = mt11;
+    if (mt11 > mt_hi)
+        mt_hi = mt11;
+    if (mtcc < mt_lo)
+        mt_lo = mtcc;
+    if (mtcc > mt_hi)
+        mt_hi = mtcc;
+
+    *brackets = (AGrid_target_mr >= mr_lo && AGrid_target_mr <= mr_hi &&
+        AGrid_target_mt >= mt_lo && AGrid_target_mt <= mt_hi);
 
     interp_mr = 0.25 * (mr00 + mr10 + mr01 + mr11);
     interp_mt = 0.25 * (mt00 + mt10 + mt01 + mt11);
@@ -181,6 +224,7 @@ static void agrid_subdivide(double u0, double u1, double v0, double v1, int dept
     int c00, c10, c01, c11, cc;
     double err;
     int need_split;
+    int brackets;
 
     agrid_make_abg(u0, v0, search, &a00, &b00, &g00);
     agrid_make_abg(u1, v0, search, &a10, &b10, &g10);
@@ -194,8 +238,8 @@ static void agrid_subdivide(double u0, double u1, double v0, double v1, int dept
     c11 = agrid_add_or_get(a11, b11, g11);
     cc = agrid_add_or_get(amm, bmm, gmm);
 
-    err = agrid_raw_interp_error(c00, c10, c01, c11, cc);
-    need_split = (depth < AGRID_MIN_DEPTH) || (err > AGRID_TOL && depth < AGRID_MAX_DEPTH);
+    err = agrid_raw_interp_error(c00, c10, c01, c11, cc, &brackets);
+    need_split = (depth < AGRID_MIN_DEPTH) || (depth < AGRID_MAX_DEPTH && (err > AGRID_TOL || brackets));
 
     if (need_split) {
         agrid_subdivide(u0, um, v0, vm, depth + 1, search);
@@ -283,6 +327,8 @@ void AGrid_Build(struct measure_type m, struct invert_type r)
     int search = r.search;
 
     AGrid_N = 0;
+    AGrid_target_mr = m.m_r;
+    AGrid_target_mt = m.m_t;
 
     if (search == FIND_AB)
         AGrid_fixed_g = r.slab.g;