+++ src/mc_test.c
@@ -37,7 +37,7 @@ int main(int argc, char **argv)
     s.a = 0.0;
     s.b = 0.0;
     s.g = 0.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -48,7 +48,7 @@ int main(int argc, char **argv)
     s.b = 0.0;
     s.g = 0.0;
     s.n_slab = 2.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -62,7 +62,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -76,7 +76,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -90,7 +90,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0.15;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -105,7 +105,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -119,7 +119,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -133,7 +133,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -147,7 +147,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0.15;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -162,7 +162,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -176,7 +176,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -190,7 +190,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -204,7 +204,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
     d_slide = 0.15;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -216,7 +216,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -226,7 +226,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -236,7 +236,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.3;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -249,7 +249,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.0;
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -257,7 +257,7 @@ int main(int argc, char **argv)
     printf("%5.4f %5.4f   %5.4f %5.4f\n", ad_uru, uru, ad_utu, utu);
 
     s.g = 0.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -265,7 +265,7 @@ int main(int argc, char **argv)
     printf("%5.4f %5.4f   %5.4f %5.4f\n", ad_uru, uru, ad_utu, utu);
 
     s.n_slab = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -275,7 +275,7 @@ int main(int argc, char **argv)
 
     s.g = 0.9;
     s.b = 4;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -291,7 +291,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     s.g = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -299,7 +299,7 @@ int main(int argc, char **argv)
     printf("%5.4f %5.4f   %5.4f %5.4f\n", ad_uru, uru, ad_utu, utu);
 
     s.g = 0.9;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -309,7 +309,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -319,7 +319,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -329,7 +329,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.3;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -344,7 +344,7 @@ int main(int argc, char **argv)
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
     s.g = 0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -352,7 +352,7 @@ int main(int argc, char **argv)
     printf("%5.4f %5.4f   %5.4f %5.4f\n", ad_uru, uru, ad_utu, utu);
 
     s.g = 0.9;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -362,7 +362,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.0;
     s.n_bottom_slide = 1.0;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -372,7 +372,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.5;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);
@@ -382,7 +382,7 @@ int main(int argc, char **argv)
     s.n_slab = 1.3;
     s.n_top_slide = 1.5;
     s.n_bottom_slide = 1.5;
-    MC_RT(s, n_photons, d, d_slide, &ur1, &ut1, &uru, &utu);
+    MC_RT(s, n_photons, d, d_slide, d_slide, &ur1, &ut1, &uru, &utu);
     RT(n_quad, &s, &ad_ur1, &ad_ut1, &ad_uru, &ad_utu);
     printf("%5.4f %5.1f %5.4f %5.4f %5.4f | ", s.a, s.b, s.g, s.n_slab, s.n_top_slide);
     printf("%5.4f %5.4f | ", d, d_slide);